requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
drf-yasg
